feat: get professionals by API

+++ src/app/api/professionals/[id]/route.ts
@@ -1,7 +1,15 @@
-import { type NextRequest, NextResponse } from "next/server"
-import { prisma } from "@/lib/db"
+import { type NextRequest, NextResponse } from "next/server";
+import { prisma } from "@/lib/db";
+import { validateApiKey } from "@/lib/api-key-utils"; // Importar a função de validação
 
 export async function GET(req: NextRequest, { params }: { params: { id: string } }) {
+  // Validar API Key
+  const apiKeyHeader = req.headers.get('Authorization');
+  const validationResult = await validateApiKey(apiKeyHeader);
+  if (!validationResult.isValid) {
+    return new NextResponse("Unauthorized", { status: 401 });
+  }
+
   try {
     const professional = await prisma.professional.findUnique({
       where: { id: params.id },
@@ -19,6 +27,13 @@ export async function GET(req: NextRequest, { params }: { params: { id: string }
 }
 
 export async function PATCH(req: NextRequest, { params }: { params: { id: string } }) {
+  // Validar API Key
+  const apiKeyHeader = req.headers.get('Authorization');
+  const validationResult = await validateApiKey(apiKeyHeader);
+  if (!validationResult.isValid) {
+    return new NextResponse("Unauthorized", { status: 401 });
+  }
+
   try {
     const body = await req.json()
     const { name, phone, company } = body
@@ -44,6 +59,13 @@ export async function PATCH(req: NextRequest, { params }: { params: { id: string
 }
 
 export async function DELETE(req: NextRequest, { params }: { params: { id: string } }) {
+  // Validar API Key
+  const apiKeyHeader = req.headers.get('Authorization');
+  const validationResult = await validateApiKey(apiKeyHeader);
+  if (!validationResult.isValid) {
+    return new NextResponse("Unauthorized", { status: 401 });
+  }
+
   try {
     await prisma.professional.delete({
       where: { id: params.id },
